+++ CHANGELOG.md
@@ -7,10 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.2] - 2020-05-22
+
+### Added
+- Field to add a custom excerpt on blog posts, if populated, it shows it instead of content
+- Added image link to the split section. Only if the secondary content is an image
+
+### Changed
+- The way multi columns set their column width. Now you select the number of columns instead of basing it on the number of columns
+
+### Fixed
+- If banner only has one slide, it removes navigation, and other slider features.
+- Block icons
+
 ## [0.5.1] - 2020-05-21
 
 ### Added
-
+- Block Category named "Improve & Grow Blocks"
 - Post option to hide sidebar
 - A fourth footer width